+++ common/src/main/kotlin/dev/emassey0135/audionavigation/PoiAnnouncements.kt
@@ -41,7 +41,7 @@ object PoiAnnouncements {
   }
   private var oldPoiList = PoiList()
   private var mutex = ReentrantLock()
-  fun announceNearbyPois(interruptSpeech: Boolean, excludePrevious: Boolean, detailed: Boolean, radius: Double, maxAnnouncements: Int, verticalLimit: Optional<Double>, includedFeatures: List<String>) {
+  fun announceNearbyPois(interruptSpeech: Boolean, excludePrevious: Boolean, detailed: Boolean, radius: Int, maxAnnouncements: Int, verticalLimit: Optional<Int>, includedFeatures: List<String>) {
     val minecraftClient = MinecraftClient.getInstance()
     val player = minecraftClient.player
     if (player==null)
@@ -65,12 +65,12 @@ object PoiAnnouncements {
   }
   fun triggerAutomaticAnnouncements() {
     val config = ClientConfig.instance!!.announcements
-    val verticalLimit = if(config.enableVerticalLimit.get()) Optional.of(config.verticalLimit.get().toDouble()) else Optional.empty()
-    announceNearbyPois(false, true, config.detailedAnnouncements.get(), config.announcementRadius.get().toDouble(), config.maxAnnouncements.get(), verticalLimit, config.includedFeatures.get())
+    val verticalLimit = if(config.enableVerticalLimit.get()) Optional.of(config.verticalLimit.get()) else Optional.empty()
+    announceNearbyPois(false, true, config.detailedAnnouncements.get(), config.announcementRadius.get(), config.maxAnnouncements.get(), verticalLimit, config.includedFeatures.get())
   }
   fun triggerManualAnnouncements() {
     val config = ClientConfig.instance!!.manualAnnouncements
-    val verticalLimit = if(config.enableVerticalLimit.get()) Optional.of(config.verticalLimit.get().toDouble()) else Optional.empty()
-    announceNearbyPois(true, false, config.detailedAnnouncements.get(), config.announcementRadius.get().toDouble(), config.maxAnnouncements.get(), verticalLimit, config.includedFeatures.get())
+    val verticalLimit = if(config.enableVerticalLimit.get()) Optional.of(config.verticalLimit.get()) else Optional.empty()
+    announceNearbyPois(true, false, config.detailedAnnouncements.get(), config.announcementRadius.get(), config.maxAnnouncements.get(), verticalLimit, config.includedFeatures.get())
   }
 }
